+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - PVC -> right click -> Go to PV
 - reveal
+- open ns obj on click on namespace
 ### Fixed
 - metadata.selfLink is deprecated